move config globals to config.py
